requirements.txt
# Core
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0

# Scraping
playwright>=1.41.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# APIs
gql[requests]>=3.5.0  # GraphQL for Product Hunt
praw>=7.7.0            # Reddit API

# AI Analysis
google-generativeai>=0.4.0  # Gemini API
groq>=0.4.0                # Groq API (Fallback)

# Google Services
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0
gspread>=6.0.0

# Email
aiosmtplib>=3.0.0

# Utilities
rich>=13.7.0           # Better console output
PyYAML>=6.0.0          # Config file parsing
tenacity>=8.2.0        # Retry logic
schedule>=1.2.0        # Scheduling

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
